+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes
+
+- Add missing context for watchdog termination events (#5242)
+
 ## 8.52.0-beta
 
 ### Features